+++ Homework/PrintingRules.Setters.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 
-namespace ObjectPrinting
+namespace Homework
 {
     internal partial class PrintingRules
     {
