+++ Homework/SerialisationContexts/SerialisationTargetConfig.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq.Expressions;
 
-namespace ObjectPrinting
+namespace Homework.SerialisationContexts
 {
     public class SerialisationTargetConfig<TOwner>
     {
