+++ Homework/SerialisationRule.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ObjectPrinting
+namespace Homework
 {
     public record SerialisationRule
     {
