+++ HomeworkTests/Person.cs
@@ -1,16 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace ObjectPrinting.Tests
+namespace HomeworkTests
 {
-    public class PersonWithList
-    {
-        public List<int> List { get; set; }
-        public int Number { get; set; } = 12378;
-        public Dictionary<string, float> Dict { get; set; }
-        public Queue<Guid> Guids { get; set; }
-    }
-    
     public class Person
     {
         public Guid Id { get; set; }
