+++ HomeworkTests/PrintIfFailure.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace HomeworkTests
+{
+    public class PrintIfFailure : Attribute
+    {
+        
+    }
+}
